feat: batch delete and edit items in cart (#TS168)

+++ src/controller/cartController.ts
@@ -2,9 +2,14 @@ import { DeleteItemDTO } from '../dto/cart/deleteItemDto';
 import { EditCartDTO } from '../dto/cart/editCartDto';
 import { GetCartDTO } from '../dto/cart/getCartDto';
 import { CartService } from '../service/cartService';
-import { IEditCartReq, IGetCartReq } from '../types/cart.type';
-import { IUserReq } from '../types/common.type';
+import {
+  DeleteCartType,
+  IDeleteCartReq,
+  IEditCartReq,
+  IGetCartReq,
+} from '../types/cart.type';
 import { CustomResponseType } from '../types/customResponseType';
+import { EditCartVo } from '../vo/cart/editCartVo';
 import { GetCartVO } from '../vo/cart/getCartVo';
 import { BaseController } from './baseController';
 
@@ -22,23 +27,36 @@ export class CartController extends BaseController {
     );
   };
 
-  public readonly deleteItem = async (req: IUserReq) => {
+  public readonly deleteItem = async (req: IDeleteCartReq) => {
     const deleteItemDto = new DeleteItemDTO(req);
-    const item = await this.cartService.deleteItem(deleteItemDto);
+
+    const { type, userId } = deleteItemDto;
+
+    if (type === DeleteCartType.all) {
+      await this.cartService.clearCart(userId);
+      return this.formatResponse(
+        CustomResponseType.OK_MESSAGE,
+        CustomResponseType.OK,
+        { items: [], errors: [] },
+      );
+    }
+
+    const carts = await this.cartService.deleteItem(deleteItemDto);
+
     return this.formatResponse(
       CustomResponseType.OK_MESSAGE,
       CustomResponseType.OK,
-      { item },
+      new EditCartVo(carts),
     );
   };
 
   public readonly editCart = async (req: IEditCartReq) => {
     const editCartDto = new EditCartDTO(req);
-    const updatedCart = await this.cartService.editCart(editCartDto);
+    const carts = await this.cartService.editCart(editCartDto);
     return this.formatResponse(
       CustomResponseType.OK_MESSAGE,
       CustomResponseType.OK,
-      { cart: updatedCart },
+      new EditCartVo(carts),
     );
   };
 }
